[MultiTarget] Extract MultiTarget to separate file
